+++ iris-mpc-utils/src/utils/fsys.rs
@@ -5,18 +5,8 @@ use serde_json;
 use std::{
     fs::File,
     io::{BufReader, BufWriter},
-    path::{Path, PathBuf},
+    path::Path,
 };
-
-/// Returns path to root directory.
-pub fn get_path_to_root() -> PathBuf {
-    Path::new(&env!("CARGO_MANIFEST_DIR").to_string()).into()
-}
-
-/// Returns path to sub-directory.
-pub fn get_path_to_subdir(name: &str) -> PathBuf {
-    get_path_to_root().join(name)
-}
 
 /// Reads binary data from a file & deserializes a domain type.
 pub fn read_bin<T: DeserializeOwned>(fpath: &Path) -> Result<T> {
@@ -52,19 +42,4 @@ pub fn write_json<T: Serialize>(data: &T, fpath: &Path) -> Result<()> {
     serde_json::to_writer(writer, &data)?;
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{get_path_to_root, get_path_to_subdir};
-
-    #[test]
-    fn test_get_path_to_root() {
-        assert!(get_path_to_root().exists());
-    }
-
-    #[test]
-    fn test_get_path_to_subdir() {
-        assert!(get_path_to_subdir("assets").exists());
-    }
 }
